move several more bento components
