+++ build-system/compile/bundles.config.bento.json
@@ -3,8 +3,7 @@
     "name": "bento-accordion",
     "version": "1.0",
     "options": {
-      "hasCss": true,
-      "npm": true
+      "hasCss": true
     }
   },
   {
@@ -25,6 +24,30 @@
   },
   {
     "name": "bento-dailymotion",
+    "version": "1.0",
+    "latestVersion": "0.1",
+    "options": {
+      "hasCss": true
+    }
+  },
+  {
+    "name": "bento-date-countdown",
+    "version": "1.0",
+    "latestVersion": "0.1",
+    "options": {
+      "hasCss": true
+    }
+  },
+  {
+    "name": "bento-date-display",
+    "version": "1.0",
+    "latestVersion": "0.1",
+    "options": {
+      "hasCss": true
+    }
+  },
+  {
+    "name": "bento-embedly-card",
     "version": "1.0",
     "latestVersion": "0.1",
     "options": {
