+++ src/bento/components/bento-date-countdown/1.0/messages.js
