+++ src/bento/components/bento-date-display/1.0/component.js
@@ -1,3 +1,5 @@
+import {getTimeZoneName} from '#bento/components/bento-date-display/format';
+
 import {getDate} from '#core/types/date';
 
 import * as Preact from '#preact';
@@ -7,8 +9,6 @@ import {useRenderer} from '#preact/component/renderer';
 import {useResourcesNotify} from '#preact/utils';
 
 import {user} from '#utils/log';
-
-import {getTimeZoneName} from '../format';
 
 /** @const {string} */
 const TAG = 'amp-date-display';
