+++ src/bento/components/bento-date-display/1.0/component.type.js
