+++ src/bento/components/bento-date-display/1.0/test/test-component.js
@@ -1,10 +1,10 @@
 import {mount} from 'enzyme';
 
+import {BentoDateDisplay} from '#bento/components/bento-date-display/1.0/component';
+
 import * as Preact from '#preact';
 
 import {user} from '#utils/log';
-
-import {BentoDateDisplay} from '../component';
 
 describes.sandboxed('DateDisplay 1.0 preact component', {}, (env) => {
   let sandbox;
