+++ src/bento/components/bento-date-display/OWNERS
@@ -1,0 +1,10 @@
+// For an explanation of the OWNERS rules and syntax, see:
+// https://github.com/ampproject/amp-github-apps/blob/main/owners/OWNERS.example
+
+{
+  rules: [
+    {
+      owners: [{name: 'robinpokorny'}, {name: 'ampproject/wg-components'}],
+    },
+  ],
+}
