+++ src/bento/components/bento-embedly-card/1.0/bento-embedly-card.js
@@ -1,2 +1,3 @@
 import {defineElement} from './web-component';
+
 defineElement();
