+++ src/bento/components/bento-embedly-card/1.0/key-base-element.js
